feat(Result): 배당금을 계산한다
